# hookd/pkg/liveplanet/api/swagger/streams_api.yaml
swagger: '2.0'
info:
  version: "1.0"
  title: Live Planet Private Streams API
  description: Live Planet Private Streams API
host: api.liveplanetstage.net
basePath: /api/private/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  Token:
    type: apiKey
    name: Authorization
    in: header
paths:
  /streams/:
    get:
      summary: Get streams list
      tags:
        - Streams
      operationId: GetStreams
      security:
        - Token: []
      responses:
        "200":
          description: success
          schema:
            type: array
            items:
              $ref: '#/definitions/Stream'
    post:
      summary: Create stream
      tags:
        - Streams
      operationId: CreateStream
      security:
        - Token: []
      parameters:
        - name: stream
          in: body
          description: stream
          required: true
          type: object
          schema:
            $ref: '#/definitions/CreateStreamRequest'
      responses:
        "201":
          description: success created
          schema:
            type: object
            $ref: '#/definitions/StreamDetail'
        "401":
          desciption: unauthorized
...
definitions:
  CreateStreamRequest:
    properties:
      name:
        type: string
      metadata:
        type: string
        description: json string with video metadata
    required:
      - name
...
  Stream:
    properties:
      id:
        type: integer
        format: int64
        description: unique id
      name:
        type: string
        description: name
      is_record:
        type: boolean
      is_processing:
        type: boolean
      video_preview_url:
        type: string
        description: absolute preview video url
    required:
      - id
      - name
      - is_record
      - is_processing
      - video_preview_url
  StreamDetail:
    properties:
      id:
        type: integer
        format: int64
        description: unique id
      name:
        type: string
        description: name
      is_record:
        type: boolean
      is_processing:
        type: boolean
      video_type:
        type: string
        description: video type
        enum:
          - MONO
          - STEREO_TB
          - STEREO_LR
      video_proj:
        type: string
        description: stream projection
        enum:
          - EQUIRECTANGULAR
          - SMOOTH_EQUIRECTANGULAR
          - CUBEMAP
      video_url:
        type: string
        description: absolute video url
      video_preview_url:
        type: string
        description: absolute preview video url
      tags:
        type: array
        description: tags list
        items:
          $ref: string
    required:
      - id
      - name
      - is_record
      - is_processing
      - video_type
      - video_proj
      - video_url
      - video_preview_url
      - tags